feat:调整代收匿名账户从列表中查询

+++ src/main/java/com/rfchina/wallet/server/service/ConfigService.java
@@ -171,7 +171,7 @@ public class ConfigService {
 	private String anonyPayerWalletIdList;
 
 	@Value("${yunst.collect.limit}")
-	private Long collectLimit;
+	private Integer collectLimit;
 
 	/**
 	 * 异常邮件通知
